+++ packages/next/src/provider/GTProvider.tsx
@@ -1,8 +1,4 @@
-import {
-  DictionaryEntry,
-  mergeDictionaries,
-  TranslationsStatus,
-} from 'gt-react/internal';
+import { DictionaryEntry, mergeDictionaries } from 'gt-react/internal';
 import { isValidElement } from 'react';
 import getI18NConfig from '../config-dir/getI18NConfig';
 import { getLocale } from '../request/getLocale';
@@ -11,11 +7,21 @@ import { Dictionary, Translations } from 'gt-react/internal';
 import { createDictionarySubsetError } from '../errors/createErrors';
 import ClientProvider from './ClientProviderWrapper';
 import { GTProviderProps } from '../utils/types';
+import { getRegion } from '../request/getRegion';
+
+/*
+Note: In normal circumstances, both _locale and _region would be at risk of causing hydration errors.
+They would be advised against as parameters of GTProvider.
+However:
+- _region is used only on the client side, and is accessed on the server purely downstream of being set as a cookie by the client
+- A disparity between _locale and the server side locale will cause the window to reload in order to set _locale as the server side locale too
+*/
 
 export default async function GTProvider({
   children,
   id: prefixId,
   locale: _locale,
+  region: _region,
 }: GTProviderProps) {
   // ---------- SETUP ---------- //
   const I18NConfig = getI18NConfig();
@@ -66,20 +72,17 @@ export default async function GTProvider({
 
   // Block until cache check resolves
   const translations = await cachedTranslationsPromise;
-  const translationsStatus: TranslationsStatus = translationRequired
-    ? I18NConfig.getCachedTranslationsStatus(locale)
-    : ({} as any);
 
   return (
     <ClientProvider
       dictionary={dictionary}
-      initialTranslations={translations}
-      initialTranslationsStatus={translationsStatus}
+      translations={translations}
       locale={locale}
       locales={I18NConfig.getLocales()}
       defaultLocale={defaultLocale}
       translationRequired={translationRequired}
       dialectTranslationRequired={dialectTranslationRequired}
+      region={_region || (await getRegion())}
       gtServicesEnabled={
         process.env._GENERALTRANSLATION_GT_SERVICES_ENABLED === 'true'
       }