+++ packages/next/src/server-dir/buildtime/T.tsx
@@ -10,12 +10,7 @@ import {
   writeChildrenAsObjects,
 } from 'gt-react/internal';
 import renderVariable from '../variables/renderVariable';
-import React from 'react';
 import { hashSource } from 'generaltranslation/id';
-
-async function Resolver({ children }: { children: React.ReactNode }) {
-  return await children;
-}
 
 /**
  * Build-time translation component that renders its children in the user's given locale.
@@ -102,11 +97,8 @@ async function T({
 
   // Get the translation entry object
   const translations = await translationsPromise;
-  const translationsStatus = translationRequired
-    ? I18NConfig.getCachedTranslationsStatus(locale)
-    : undefined;
+
   let translationEntry = translations?.[id || ''];
-  let translationsStatusEntry = translationsStatus?.[id || ''];
 
   let childrenAsObjects;
   let hash;
@@ -124,7 +116,6 @@ async function T({
         dataFormat: 'JSX',
       });
     translationEntry = translations?.[hash];
-    translationsStatusEntry = translationsStatus?.[hash];
   }
 
   // ----- RENDERING FUNCTION #2: RENDER TRANSLATED CONTENT ----- //
@@ -141,19 +132,19 @@ async function T({
   // ----- RENDER CACHED TRANSLATIONS ----- //
 
   // if we have a cached translation, render it
-  if (translationsStatusEntry?.status === 'success') {
-    return renderTranslation(translationEntry);
+  if (translationEntry) {
+    return <Suspense>{renderTranslation(translationEntry)}</Suspense>;
   }
 
-  if (translationsStatusEntry?.status === 'error') {
+  if (
+    translationEntry === null || // If there was an error
+    !I18NConfig.isDevelopmentApiEnabled() // Development translation disabled
+  ) {
     return renderDefault();
   }
 
   // ----- TRANSLATE ON DEMAND ----- //
-  // dev only
-
-  // Since this is the buildtime translation component <T>, this is dev-only
-  if (!I18NConfig.isDevelopmentApiEnabled()) return renderDefault();
+  // Since this is the buildtime translation component <T>, everything below is dev-only
 
   // Get render settings
   const renderSettings = I18NConfig.getRenderSettings();
@@ -203,7 +194,7 @@ async function T({
 
   return (
     <Suspense key={locale} fallback={loadingFallback}>
-      <Resolver children={translationPromise} />
+      {translationPromise}
     </Suspense>
   );
 }