+++ packages/react/src/provider/ClientProvider.tsx
@@ -4,17 +4,20 @@ import { useEffect, useMemo, useState } from 'react';
 import { determineLocale, GT } from 'generaltranslation';
 import { GTContext } from './GTContext';
 import { ClientProviderProps } from '../types/config';
-import { TranslationsStatus, Translations } from '../types/types';
+import { Translations } from '../types/types';
 import useRuntimeTranslation from './hooks/useRuntimeTranslation';
 import useCreateInternalUseGTFunction from './hooks/useCreateInternalUseGTFunction';
 import useCreateInternalUseTranslationsFunction from './hooks/useCreateInternalUseTranslationsFunction';
-import { defaultLocaleCookieName } from '../utils/cookies';
+import {
+  defaultLocaleCookieName,
+  defaultRegionCookieName,
+} from '../utils/cookies';
+
 // meant to be used inside the server-side <GTProvider>
 export default function ClientProvider({
   children,
   dictionary,
-  initialTranslations,
-  initialTranslationsStatus,
+  translations: _translations,
   locale: _locale,
   region: _region,
   _versionId,
@@ -29,18 +32,15 @@ export default function ClientProvider({
   runtimeTranslationEnabled,
   resetLocaleCookieName,
   localeCookieName = defaultLocaleCookieName,
+  regionCookieName = defaultRegionCookieName,
   customMapping,
 }: ClientProviderProps): React.JSX.Element {
   // ----- TRANSLATIONS STATE ----- //
 
   const [translations, setTranslations] = useState<Translations | null>(
-    devApiKey ? null : initialTranslations
+    // devApiKey ? null : _translations
+    _translations // likely to induce hydration error
   );
-
-  const [translationsStatus, setTranslationsStatus] =
-    useState<TranslationsStatus | null>(
-      devApiKey ? null : initialTranslationsStatus
-    );
 
   // ----- LOCALE STATE ----- //
 
@@ -49,15 +49,20 @@ export default function ClientProvider({
     _locale ? determineLocale(_locale, locales) || '' : ''
   );
 
-  // Monitor for changes in _locale parameter
-  useEffect(() => {
-    const newLocale = _locale ? determineLocale(_locale, locales) || '' : '';
-    if (newLocale !== locale) {
-      _setLocale(newLocale);
-    }
-  }, [_locale, locales]);
+  // Set the locale via cookies and refresh the page to reload server-side. Make sure the language is supported.
+  const setLocale = (newLocale: string): void => {
+    // validate locale
+    newLocale = determineLocale(newLocale, locales) || locale || defaultLocale;
+    // persist locale
+    document.cookie = `${localeCookieName}=${newLocale};path=/`;
+    document.cookie = `${resetLocaleCookieName}=true;path=/`;
+    // set locale
+    _setLocale(newLocale);
+    // re-render server components
+    window.location.reload();
+  };
 
-  // Check for an invalid cookie
+  // Check for an invalid cookie and update it
   useEffect(() => {
     const cookieLocale = document.cookie
       .split('; ')
@@ -68,22 +73,20 @@ export default function ClientProvider({
     }
   }, [locale, localeCookieName]);
 
-  // Set the locale via cookies and refresh the page to reload server-side. Make sure the language is supported.
-  const setLocale = (newLocale: string): void => {
-    // validate locale
-    newLocale = determineLocale(newLocale, locales) || locale || defaultLocale;
+  // ----- REGION STATE ----- //
 
-    // persist locale
-    document.cookie = `${localeCookieName}=${newLocale};path=/`;
-    document.cookie = `${resetLocaleCookieName}=true;path=/`;
+  // Set region state
+  const [region, _setRegion] = useState(_region);
 
-    // set locale
-    _setLocale(newLocale);
-
-    // re-render client components
-    window.location.reload();
+  // Set the region via cookies. No page reload needed.
+  const setRegion = (newRegion: string | undefined): void => {
+    // persist region
+    document.cookie = `${regionCookieName}=${newRegion || ''};path=/`;
+    // set region
+    _setRegion(newRegion);
   };
 
+<<<<<<< HEAD
   const [region, setRegion] = useState(_region);
 
   // ----- GT SETUP ----- //
@@ -106,45 +109,54 @@ export default function ClientProvider({
   // ---------- TRANSLATION LIFECYCLE ---------- //
 
   // Fetch additional translations and queue them for merging
+=======
+  // Check for an invalid cookie and update it
+>>>>>>> a/t
   useEffect(() => {
-    setTranslations((prev) => ({ ...initialTranslations, ...prev }));
-    setTranslationsStatus((prev) => ({
-      ...Object.keys(initialTranslations).reduce(
-        (acc: TranslationsStatus, hash) => {
-          acc[hash] = {
-            status: 'success',
-          };
-          return acc;
-        },
-        {}
-      ),
-      ...prev,
-    }));
-  }, [initialTranslations]);
+    const cookieRegion = document.cookie
+      .split('; ')
+      .find((row) => row.startsWith(`${regionCookieName}=`))
+      ?.split('=')[1];
+    if (region && cookieRegion && cookieRegion !== region) {
+      document.cookie = `${regionCookieName}=;path=/`;
+    }
+  }, [region, regionCookieName]);
+
+  // ----- GT SETUP ----- //
+
+  // Define the GT instance
+  // Used for custom mapping and as a driver for the runtime translation
+  const gt = useMemo(
+    () =>
+      new GT({
+        devApiKey,
+        sourceLocale: defaultLocale,
+        targetLocale: locale,
+        projectId,
+        baseUrl: runtimeUrl || undefined,
+        customMapping,
+      }),
+    [devApiKey, defaultLocale, locale, projectId, runtimeUrl, customMapping]
+  );
 
   // ---------- TRANSLATION METHODS ---------- //
 
-  const {
-    registerIcuForTranslation,
-    registerJsxForTranslation,
-    registerI18nextForTranslation,
-  } = useRuntimeTranslation({
-    gt,
-    locale: locale,
-    versionId: _versionId,
-    runtimeUrl,
-    setTranslations,
-    setTranslationsStatus,
-    defaultLocale,
-    renderSettings,
-    runtimeTranslationEnabled,
-  });
+  const { registerIcuForTranslation, registerJsxForTranslation } =
+    useRuntimeTranslation({
+      gt,
+      locale: locale,
+      versionId: _versionId,
+      runtimeUrl,
+      setTranslations,
+      defaultLocale,
+      renderSettings,
+      runtimeTranslationEnabled,
+    });
 
   // ---------- USE GT() TRANSLATION ---------- //
 
   const _internalUseGTFunction = useCreateInternalUseGTFunction(
     translations,
-    translationsStatus,
     locale,
     defaultLocale,
     translationRequired,
@@ -160,7 +172,6 @@ export default function ClientProvider({
     useCreateInternalUseTranslationsFunction(
       dictionary,
       translations,
-      translationsStatus,
       locale,
       defaultLocale,
       translationRequired,
@@ -180,7 +191,6 @@ export default function ClientProvider({
       value={{
         gt,
         registerIcuForTranslation,
-        registerI18nextForTranslation,
         registerJsxForTranslation,
         setLocale,
         _internalUseGTFunction,
@@ -191,7 +201,6 @@ export default function ClientProvider({
         region,
         setRegion,
         translations,
-        translationsStatus: translationsStatus,
         translationRequired,
         dialectTranslationRequired,
         renderSettings,